\name{COPDSexualDimorphism-package}
\alias{COPDSexualDimorphism-package}
\alias{COPDSexualDimorphism}
\docType{package}
\title{
Data for Sexual Dimorphic and COPD Differential Analysis
}
\description{
Data package to accompany the Package COPDSexualDimorphism.
}
\details{
\tabular{ll}{
Package: \tab COPDSexualDimorphism.data\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-14\cr
License: \tab LGPL-2.1\cr
}
Along with the method package COPDSexualDimorphism, this data package is to accompany the paper "Integrative Genomics of Sexual Dimorphism in COPD" by Sathirapongsasuti et al.
}
\author{
J Fah Sathirapongsasuti

Maintainer: J Fah Sathirapongsasuti <fah@cs.stanford.edu>
}
\references{
Sathirapongsasuti JF, Glass K, Huttenhower C, Quackenbush J, DeMeo DL. Integrative
Genomics of Sexual Dimorphism in COPD. (In Prep).

www.lung-genomics.org
}
\keyword{ package }
\keyword{ COPD }
\keyword{ sexual dimorphism }
\keyword{ SDCD }
\examples{
data(lgrc.expr.meta)
data(lgrc.expr)
data(lgrc.genes)
}
